requirements.txt
python-snap7==2.0.0
tabulate==0.9.0
